Billboards
added billboards to enemies

+++ Assets/Scripts/DestroyAfterTime.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class DestroyAfterTime : MonoBehaviour
 {
-    public int destroyTime = 3;
+    public float destroyTime = 3;
     void Awake()
     {
         Destroy(this.gameObject, destroyTime);
